zod validation schema types defined and tested

+++ app/actions/actions.ts
@@ -4,9 +4,9 @@ import { revalidatePath } from "next/cache";
 import prisma from "@/lib/db";
 import { Pet } from "@prisma/client";
 import { PetEssentials } from "@/lib/types";
-import { FormSchema } from "@/lib/validation";
+import { FormSchema, IndivdiualPetId } from "@/lib/validation";
 
-export async function addPet(formData: PetEssentials) {
+export async function addPet(formData: unknown) {
 	const { success, data } = FormSchema.safeParse(formData);
 
 	if (success) {
@@ -14,6 +14,7 @@ export async function addPet(formData: PetEssentials) {
 			await prisma.pet.create({
 				data: data,
 			});
+			console.log(`successfully added the pet`);
 			revalidatePath("/app", "layout");
 		} catch (error: any) {
 			console.log(error);
@@ -22,22 +23,26 @@ export async function addPet(formData: PetEssentials) {
 			};
 		}
 	} else {
-		console.log("schema is invalid");
+		return {
+			message: "Schema is invalid.",
+		};
 	}
 }
 
-export async function editPet(formData: PetEssentials, selectedId: Pet) {
-	const { success, data } = FormSchema.safeParse(formData);
+export async function editPet(formData: unknown, selectedId: unknown) {
+	const validatedForm = FormSchema.safeParse(formData);
+	const validatedId = IndivdiualPetId.safeParse(selectedId);
 
 	// if validation is success go into the if block
-	if (success) {
+	if (validatedForm.success && validatedId.success) {
 		try {
 			await prisma.pet.update({
 				where: {
-					id: selectedId?.id,
+					id: validatedId.data,
 				},
-				data: data,
+				data: validatedForm.data,
 			});
+			console.log(`successfully udpated the pet`);
 			revalidatePath("/app", "layout");
 		} catch (error: any) {
 			return {
@@ -45,21 +50,32 @@ export async function editPet(formData: PetEssentials, selectedId: Pet) {
 			};
 		}
 	} else {
-		console.log("schema is invalid");
+		return {
+			message: "Schema is invalid.",
+		};
 	}
 }
 
-export async function deletePet(petId: string) {
-	try {
-		await prisma.pet.delete({
-			where: {
-				id: petId,
-			},
-		});
-		revalidatePath("/app", "layout");
-	} catch (error) {
+export async function deletePet(petId: unknown) {
+	const validatedId = IndivdiualPetId.safeParse(petId);
+
+	if (validatedId.success) {
+		try {
+			await prisma.pet.delete({
+				where: {
+					id: validatedId.data,
+				},
+			});
+			console.log(`successfully deleted the pet`);
+			revalidatePath("/app", "layout");
+		} catch (error) {
+			return {
+				message: "Could not delete pet.",
+			};
+		}
+	} else {
 		return {
-			message: "Could not delete pet.",
+			message: "PetId is invalid",
 		};
 	}
 }
